pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-reader-cli"
version = "0.1.0"
description = "CLI for reading, managing, and basic analysis of PDFs"
readme = "README.md"
authors = [
  { name = "Your Name" }
]
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
  "typer>=0.12",
  "rich>=13.9",
  "pypdf>=4.3",
  "pdfplumber>=0.11",
  "PyQt5>=5.15",
  "pdf2image>=1.16",
  "pymupdf>=1.24",
  "pillow>=10.0"
]

[project.scripts]
pdfreader = "pdf_reader.cli:main"
pdfreader-gui = "pdf_reader.gui:main"
pdfreader-web = "streamlit.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
